offloaded the actual database string gen. to a new private module
